requirements.txt
asgiref==3.7.2
crispy-bootstrap5==2023.10
dj-database-url==2.1.0
Django==5.0.1
django-crispy-forms==2.1
joblib==1.3.2
mysql-connector-python==8.3.0
mysqlclient==2.2.1
numpy==1.26.3
pandas==2.2.0
pycryptodomex==3.20.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.4
scikit-learn==1.2.2
scipy==1.12.0
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.2.0
typing_extensions==4.9.0
tzdata==2023.4
